Add Deltalake-Spark3 Demo
The nessiedemo-library code for the Delta-Spark demo uses a very similar approach as
the existing Iceberg-Spark nessiedemo-code. So I've refactored the related nessiedemo-library
code to a base class that concrete classes for Iceberg-Spark and Delta-Spark use.

The existing notebook `nessie-iceberg-demo-nba.ipynb` has not been renamed, as that
would break existing links to that demo.

This PR also removes the `TypeDef/T` code.

Configuration file and test-code for Delta-Spark2 is there, but disabled, because there
are issues with Delta < 0.8.