Fix mybinder
Even after #395 the demos do not work on mybinder.org. After some investigation it turned out that some tweaks are to the binder Dockerfile are required.

This changes fixes a warning that the NB_USER and NB_UID args were not consumed. Those are now consumed - verifying the non-root user name and ID are the same as the ones used in the [base image](https://github.com/projectnessie/nessie-demos/pkgs/container/nessie-binder-demos).

The actual fix is the ENTRYPOINT in the binder Dockerfile, which seems to be required, because binder passes custom arguments to jupyter.

See https://mybinder.org/v2/gh/snazy/nessie-demos/fix-mybinder?labpath=notebooks%2Fnessie-iceberg-demo-nba.ipynb